Foi feita alteraçoes no Ranking e na Avaliaçao

+++ PFC.Model/Model/Avaliacao.cs
@@ -12,8 +12,8 @@ namespace PFC.Model
         public int idAvaliacao { get; set; }
         public int  idUsuario { get; set; }
         public int idTopico { get; set; }
-        public int pontos { get; set; }
-        public int mediaPontos { get; set; }
+        public float pontos { get; set; }
+        public float mediaPontos { get; set; }
         public int contarLike { get; set; }
         public int contarDeslike { get; set; }
     }
